Add e2e test for configurable pod resolv.conf
